Split out Priority, Project, ProjectSlugHistory & ProjectQuerySet
